+++ mws/__init__.py
@@ -1,4 +1,20 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
 
-from .mws import *  # noqa: F401, F403
+from .mws import MWS, MWSError
+from .apis import Feeds, Finances, InboundShipments, Inventory, OffAmazonPayments, Orders,\
+                  Products, Recommendations, Reports, Sellers
+__all__ = [
+    'Feeds',
+    'Finances',
+    'InboundShipments',
+    'Inventory',
+    'MWS',
+    'MWSError',
+    'OffAmazonPayments',
+    'Orders',
+    'Products',
+    'Recommendations',
+    'Reports',
+    'Sellers',
+]
