+++ mws/apis/__init__.py
@@ -1,0 +1,25 @@
+from .feeds import Feeds
+from .finances import Finances
+from .inbound_shipments import InboundShipments
+from .inventory import Inventory
+from .merchant_fulfillment import MerchantFulfillment
+from .offamazonpayments import OffAmazonPayments
+from .orders import Orders
+from .products import Products
+from .recommendations import Recommendations
+from .reports import Reports
+from .sellers import Sellers
+
+__all__ = [
+    'Feeds',
+    'Finances',
+    'InboundShipments',
+    'Inventory',
+    'MerchantFulfillment',
+    'OffAmazonPayments',
+    'Orders',
+    'Products',
+    'Recommendations',
+    'Reports',
+    'Sellers',
+]
