+++ mws/apis/products.py
@@ -1,0 +1,130 @@
+"""
+Amazon MWS Products API
+"""
+from __future__ import absolute_import
+# import warnings
+
+from ..mws import MWS
+from .. import utils
+
+
+class Products(MWS):
+    """
+    Amazon MWS Products API
+    """
+    URI = '/Products/2011-10-01'
+    VERSION = '2011-10-01'
+    NAMESPACE = '{http://mws.amazonservices.com/schema/Products/2011-10-01}'
+    # NEXT_TOKEN_OPERATIONS = []
+
+    def list_matching_products(self, marketplaceid, query, contextid=None):
+        """
+        Returns a list of products and their attributes, ordered by
+        relevancy, based on a search query that you specify.
+        Your search query can be a phrase that describes the product
+        or it can be a product identifier such as a UPC, EAN, ISBN, or JAN.
+        """
+        data = dict(Action='ListMatchingProducts',
+                    MarketplaceId=marketplaceid,
+                    Query=query,
+                    QueryContextId=contextid)
+        return self.make_request(data)
+
+    def get_matching_product(self, marketplaceid, asins):
+        """
+        Returns a list of products and their attributes, based on a list of
+        ASIN values that you specify.
+        """
+        data = dict(Action='GetMatchingProduct', MarketplaceId=marketplaceid)
+        data.update(utils.enumerate_param('ASINList.ASIN.', asins))
+        return self.make_request(data)
+
+    def get_matching_product_for_id(self, marketplaceid, type_, ids):
+        """
+        Returns a list of products and their attributes, based on a list of
+        product identifier values (ASIN, SellerSKU, UPC, EAN, ISBN, GCID  and JAN)
+        The identifier type is case sensitive.
+        Added in Fourth Release, API version 2011-10-01
+        """
+        data = dict(Action='GetMatchingProductForId',
+                    MarketplaceId=marketplaceid,
+                    IdType=type_)
+
+        data.update(utils.enumerate_param('IdList.Id.', ids))
+        return self.make_request(data)
+
+    def get_competitive_pricing_for_sku(self, marketplaceid, skus):
+        """
+        Returns the current competitive pricing of a product,
+        based on the SellerSKU and MarketplaceId that you specify.
+        """
+        data = dict(Action='GetCompetitivePricingForSKU', MarketplaceId=marketplaceid)
+        data.update(utils.enumerate_param('SellerSKUList.SellerSKU.', skus))
+        return self.make_request(data)
+
+    def get_competitive_pricing_for_asin(self, marketplaceid, asins):
+        """
+        Returns the current competitive pricing of a product,
+        based on the ASIN and MarketplaceId that you specify.
+        """
+        data = dict(Action='GetCompetitivePricingForASIN', MarketplaceId=marketplaceid)
+        data.update(utils.enumerate_param('ASINList.ASIN.', asins))
+        return self.make_request(data)
+
+    def get_lowest_offer_listings_for_sku(self, marketplaceid, skus, condition="Any", excludeme="False"):
+        data = dict(Action='GetLowestOfferListingsForSKU',
+                    MarketplaceId=marketplaceid,
+                    ItemCondition=condition,
+                    ExcludeMe=excludeme)
+        data.update(utils.enumerate_param('SellerSKUList.SellerSKU.', skus))
+        return self.make_request(data)
+
+    def get_lowest_offer_listings_for_asin(self, marketplaceid, asins, condition="Any", excludeme="False"):
+        data = dict(Action='GetLowestOfferListingsForASIN',
+                    MarketplaceId=marketplaceid,
+                    ItemCondition=condition,
+                    ExcludeMe=excludeme)
+        data.update(utils.enumerate_param('ASINList.ASIN.', asins))
+        return self.make_request(data)
+
+    def get_lowest_priced_offers_for_sku(self, marketplaceid, sku, condition="New", excludeme="False"):
+        data = dict(Action='GetLowestPricedOffersForSKU',
+                    MarketplaceId=marketplaceid,
+                    SellerSKU=sku,
+                    ItemCondition=condition,
+                    ExcludeMe=excludeme)
+        return self.make_request(data)
+
+    def get_lowest_priced_offers_for_asin(self, marketplaceid, asin, condition="New", excludeme="False"):
+        data = dict(Action='GetLowestPricedOffersForASIN',
+                    MarketplaceId=marketplaceid,
+                    ASIN=asin,
+                    ItemCondition=condition,
+                    ExcludeMe=excludeme)
+        return self.make_request(data)
+
+    def get_product_categories_for_sku(self, marketplaceid, sku):
+        data = dict(Action='GetProductCategoriesForSKU',
+                    MarketplaceId=marketplaceid,
+                    SellerSKU=sku)
+        return self.make_request(data)
+
+    def get_product_categories_for_asin(self, marketplaceid, asin):
+        data = dict(Action='GetProductCategoriesForASIN',
+                    MarketplaceId=marketplaceid,
+                    ASIN=asin)
+        return self.make_request(data)
+
+    def get_my_price_for_sku(self, marketplaceid, skus, condition=None):
+        data = dict(Action='GetMyPriceForSKU',
+                    MarketplaceId=marketplaceid,
+                    ItemCondition=condition)
+        data.update(utils.enumerate_param('SellerSKUList.SellerSKU.', skus))
+        return self.make_request(data)
+
+    def get_my_price_for_asin(self, marketplaceid, asins, condition=None):
+        data = dict(Action='GetMyPriceForASIN',
+                    MarketplaceId=marketplaceid,
+                    ItemCondition=condition)
+        data.update(utils.enumerate_param('ASINList.ASIN.', asins))
+        return self.make_request(data)
